Major Refactor
The biggest change here is the incorporation of a package hierarchy and
the development of the OCMhap package. Creating this package should make
distribution and (eventually) installation a little easier. Also added
controllers for the main page and data import page to remove some logic
from the GUI itself.

TODOs include implementing data import, both from a file and manually, as
well as implementation of the visualizations/analysis themselves.
Finally, further features and interfaces should be decided upon and
discussed.